call-message-extraction/requirements.txt
pillow
pytesseract
configs